+++ agat/model/__init__.py
@@ -5,7 +5,7 @@ Created on Tue Aug 15 11:41:20 2023
 @author: ZHANG Jun
 """
 
-from .GatEnergyModel import *
-from .GatForceModel import *
-from .SingleGatLayer import *
-from .ModelFit import *
+from .GatEnergyModel import EnergyGat
+from .GatForceModel import ForceGat
+from .SingleGatLayer import GATLayer
+from .ModelFit import Train